tail optimization: remove subprocess overhead
Subprocess tailing with lines require many read and seeks in order
to find the position of starting line, and then it is a simple
copy-to-stdout. The former is extremely unnecessary, since we do not
require an accurate starting line.

Seeking to a similar position with SEEK_END will be much faster, and
easier to implement. Now that is is implemented within Python, we no
longer have the overhead of process initialization and inter-process
communication.

The last 10000 lines has 3679657 bytes. Rounding up, we change each
line to 400 bytes.